+++ ui-design-lib/src/main/java/co/electriccoin/zcash/ui/design/theme/ZcashTheme.kt
@@ -27,22 +27,14 @@ fun ZcashTheme(
     content: @Composable () -> Unit
 ) {
     val useDarkMode = forceDarkMode || isSystemInDarkTheme()
+    val baseColors = if (useDarkMode) DarkColorPalette else LightColorPalette
+    val extendedColors = if (useDarkMode) DarkExtendedColorPalette else LightExtendedColorPalette
+    val zashiColors = if (useDarkMode) DarkZashiColorPalette else LightZashiColorPalette
 
-    val baseColors =
-        if (useDarkMode) {
-            DarkColorPalette
-        } else {
-            LightColorPalette
-        }
-
-    val extendedColors =
-        if (useDarkMode) {
-            DarkExtendedColorPalette
-        } else {
-            LightExtendedColorPalette
-        }
-
-    CompositionLocalProvider(LocalExtendedColors provides extendedColors) {
+    CompositionLocalProvider(
+        LocalExtendedColors provides extendedColors,
+        LocalZashiColors provides zashiColors,
+    ) {
         ProvideDimens {
             MaterialTheme(
                 colorScheme = baseColors,
@@ -72,4 +64,8 @@ object ZcashTheme {
     val dimens: Dimens
         @Composable
         get() = localDimens.current
+
+    val zashiColors: ZashiColors
+        @Composable
+        get() = LocalZashiColors.current
 }