+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/fixture/BalanceStateFixture.kt
@@ -2,6 +2,7 @@ package co.electriccoin.zcash.ui.fixture
 
 import cash.z.ecc.android.sdk.model.Zatoshi
 import co.electriccoin.zcash.ui.common.compose.BalanceState
+import co.electriccoin.zcash.ui.common.wallet.ExchangeRateState
 
 object BalanceStateFixture {
     private const val BALANCE_VALUE = 0L
@@ -11,9 +12,11 @@ object BalanceStateFixture {
 
     fun new(
         totalBalance: Zatoshi = TOTAL_BALANCE,
-        spendableBalance: Zatoshi = SPENDABLE_BALANCE
+        spendableBalance: Zatoshi = SPENDABLE_BALANCE,
+        exchangeRate: ExchangeRateState = ObserveFiatCurrencyResultFixture.new()
     ) = BalanceState.Available(
         totalBalance = totalBalance,
-        spendableBalance = spendableBalance
+        spendableBalance = spendableBalance,
+        exchangeRate = exchangeRate
     )
 }