+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/fixture/ExchangeRateStateFixture.kt
@@ -1,0 +1,9 @@
+package co.electriccoin.zcash.ui.fixture
+
+import co.electriccoin.zcash.ui.common.wallet.ExchangeRateState
+
+object ExchangeRateStateFixture {
+    val STATE = ExchangeRateState.OptedOut
+
+    fun new(state: ExchangeRateState = STATE) = state
+}