+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/exchangerate/ZashiTextButton.kt
@@ -1,0 +1,46 @@
+package co.electriccoin.zcash.ui.screen.exchangerate
+
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.ButtonColors
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import co.electriccoin.zcash.ui.design.theme.ZcashTheme
+
+@Composable
+internal fun ZashiTextButton(
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+    enabled: Boolean = true,
+    colors: ButtonColors = ZashiTextButtonDefaults.textButtonColors(),
+    content: @Composable RowScope.() -> Unit
+) {
+    TextButton(
+        onClick = onClick,
+        modifier = modifier,
+        shape = RoundedCornerShape(12.dp),
+        enabled = enabled,
+        colors = colors,
+        content = content
+    )
+}
+
+object ZashiTextButtonDefaults {
+    @Composable
+    fun textButtonColors(
+        containerColor: Color = Color.Unspecified,
+        contentColor: Color = ZcashTheme.zashiColors.btnTextFg,
+        disabledContainerColor: Color = Color.Unspecified,
+        disabledContentColor: Color = Color.Unspecified,
+    ): ButtonColors =
+        ButtonDefaults.textButtonColors(
+            containerColor = containerColor,
+            contentColor = contentColor,
+            disabledContainerColor = disabledContainerColor,
+            disabledContentColor = disabledContentColor
+        )
+}